services/tts-service/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
openai==1.58.0
pydantic==2.10.0
google-cloud-storage==2.19.0
